implement money withdrawal during transaction

+++ back-end/Common/Common.BankingAccounts/Common/AccountBalanceDto.cs
@@ -2,4 +2,4 @@ namespace Common.Accounts.Common;
 
 public record AccountBalanceDto(
     Guid AccountId, 
-    ulong Balance);
+    double Balance);
