+++ back-end/Services/Checkings/Checkings.API/Models/Domain/CheckingAccount.cs
@@ -7,7 +7,7 @@ public class CheckingAccount
     public Guid Id { get; set; }
     public required string OwnerEmail { get; set; }
     public required Guid AccountNumber { get; set; }
-    public required ulong Balance { get; set; }
+    public required double Balance { get; set; }
     public required DateTime CreatedAt { get; set; }
     public required AccountStatus Status { get; set; }
     public required string Currency { get; set; }
