+++ back-end/Services/Savings/Savings.API/Models/Domain/SavingAccount.cs
@@ -8,7 +8,7 @@ public abstract class SavingAccount
     public Guid Id { get; set; }
     public required string EmailOwner { get; set; }
     public required Guid AccountNumber { get; set; }
-    public required ulong Balance { get; set; }
+    public required double Balance { get; set; }
     public required DateTime CreatedAt { get; set; }
     public required AccountStatus Status { get; set; }
     public required string Currency { get; set; }
